feat: update `quantity`, `current_price`, and `trading_cost` types in `SymbolPnL` and `SymbolExposure` to `double` in protobuf definition

+++ rust/src/kdo.v1.fund.rs
@@ -97,14 +97,14 @@ pub struct SymbolPnL {
     #[prost(string, tag="1")]
     pub symbol: ::prost::alloc::string::String,
     /// 보유 수량
-    #[prost(int64, tag="2")]
-    pub quantity: i64,
+    #[prost(double, tag="2")]
+    pub quantity: f64,
     /// 평균 진입가
     #[prost(double, tag="3")]
     pub average_entry_price: f64,
     /// 현재가
-    #[prost(int64, tag="4")]
-    pub current_price: i64,
+    #[prost(double, tag="4")]
+    pub current_price: f64,
     /// 미실현 손익
     #[prost(int64, tag="5")]
     pub unrealized_pnl: i64,
@@ -112,8 +112,8 @@ pub struct SymbolPnL {
     #[prost(int64, tag="6")]
     pub realized_pnl: i64,
     /// 거래 비용
-    #[prost(int64, tag="7")]
-    pub trading_cost: i64,
+    #[prost(double, tag="7")]
+    pub trading_cost: f64,
 }
 /// 펀드 익스포저(Exposure) 관리
 #[allow(clippy::derive_partial_eq_without_eq)]
@@ -133,11 +133,11 @@ pub struct SymbolExposure {
     #[prost(string, tag="1")]
     pub symbol: ::prost::alloc::string::String,
     /// 보유 수량
-    #[prost(float, tag="2")]
-    pub quantity: f32,
+    #[prost(double, tag="2")]
+    pub quantity: f64,
     /// 현재가
-    #[prost(int64, tag="3")]
-    pub current_price: i64,
+    #[prost(double, tag="3")]
+    pub current_price: f64,
     /// 익스포저
     #[prost(int64, tag="4")]
     pub exposure: i64,
